+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntityTableConditionModelTest.java
@@ -55,11 +55,11 @@ public class DefaultEntityTableConditionModelTest {
 		assertEquals(Employee.TYPE, conditionModel.entityType());
 		assertEquals(10, conditionModel.conditionModels().size());
 
-		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 
-		assertFalse(conditionModel.enabled());
+		assertFalse(conditionModel.enabled().get());
 		conditionModel.conditionModel(Employee.DEPARTMENT_FK).enabled().set(true);
-		assertTrue(conditionModel.enabled());
+		assertTrue(conditionModel.enabled().get());
 	}
 
 	@Test
@@ -86,27 +86,27 @@ public class DefaultEntityTableConditionModelTest {
 	@Test
 	void setEqualOperand() throws DatabaseException {
 		Entity sales = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("SALES"));
-		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 		boolean searchStateChanged = conditionModel.setEqualOperand(Employee.DEPARTMENT_FK, sales);
 		assertTrue(searchStateChanged);
-		assertTrue(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertTrue(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 		ColumnConditionModel<Attribute<?>, Entity> deptModel =
 						conditionModel.attributeModel(Employee.DEPARTMENT_FK);
 		assertSame(deptModel.operands().equal().get(), sales);
 		assertThrows(NullPointerException.class, () -> conditionModel.setEqualOperand(null, sales));
 		searchStateChanged = conditionModel.setEqualOperand(Employee.DEPARTMENT_FK, null);
 		assertTrue(searchStateChanged);
-		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 	}
 
 	@Test
 	void setInOperands() throws DatabaseException {
 		Entity sales = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("SALES"));
 		Entity accounting = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("ACCOUNTING"));
-		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 		boolean searchStateChanged = conditionModel.setInOperands(Employee.DEPARTMENT_FK, asList(sales, accounting));
 		assertTrue(searchStateChanged);
-		assertTrue(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertTrue(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 		ColumnConditionModel<Attribute<?>, Entity> deptModel =
 						conditionModel.attributeModel(Employee.DEPARTMENT_FK);
 		assertTrue(deptModel.operands().in().get().contains(sales));
@@ -115,25 +115,25 @@ public class DefaultEntityTableConditionModelTest {
 		assertThrows(NullPointerException.class, () -> conditionModel.setInOperands(null, emptyList()));
 		searchStateChanged = conditionModel.setInOperands(Employee.DEPARTMENT_FK, emptyList());
 		assertTrue(searchStateChanged);
-		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 	}
 
 	@Test
 	void clearColumnConditionModels() throws DatabaseException {
 		Entity sales = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("SALES"));
 		Entity accounting = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("ACCOUNTING"));
-		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 		conditionModel.setInOperands(Employee.DEPARTMENT_FK, asList(sales, accounting));
-		assertTrue(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertTrue(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 		conditionModel.clear();
-		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 	}
 
 	@Test
 	void condition() throws DatabaseException {
 		Entity sales = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("SALES"));
 		Entity accounting = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("ACCOUNTING"));
-		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK));
+		assertFalse(conditionModel.enabled(Employee.DEPARTMENT_FK).get());
 		conditionModel.setInOperands(Employee.DEPARTMENT_FK, asList(sales, accounting));
 		ColumnConditionModel<?, String> nameConditionModel = conditionModel.attributeModel(Employee.NAME);
 		nameConditionModel.operands().equal().set("SCOTT");