+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/FilterComboBoxModel.java
@@ -37,7 +37,7 @@ import static java.util.Objects.requireNonNull;
  * A combo box model based on {@link FilterModel}.
  * @param <T> the type of values in this combo box model
  * @see #filterComboBoxModel()
- * @see #visiblePredicate()
+ * @see Items#visiblePredicate()
  */
 public interface FilterComboBoxModel<T> extends FilterModel<T>, ComboBoxModel<T> {
 
@@ -59,7 +59,7 @@ public interface FilterComboBoxModel<T> extends FilterModel<T>, ComboBoxModel<T>
 	 * Note that if the item does not fulfill the visible predicate, it will be filtered right away.
 	 * @param item the item to add
 	 * @throws IllegalArgumentException in case the item fails validation
-	 * @see #visiblePredicate()
+	 * @see Items#visiblePredicate()
 	 */
 	void addItem(T item);
 
@@ -137,7 +137,7 @@ public interface FilterComboBoxModel<T> extends FilterModel<T>, ComboBoxModel<T>
 	 * which is not currently visible in the model
 	 * This is false by default.
 	 * @return the {@link State} controlling whether filtering affects the selected item
-	 * @see #visiblePredicate()
+	 * @see Items#visiblePredicate()
 	 */
 	State filterSelectedItem();