+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTableCellRendererBuilder.java
@@ -71,8 +71,8 @@ final class EntityTableCellRendererBuilder extends DefaultFilterTableCellRendere
 		@Override
 		protected Color backgroundColorShaded(FilterTableModel<?, Attribute<?>> tableModel, int row,
 																					Attribute<?> identifier, Color cellBackgroundColor) {
-			boolean conditionEnabled = ((SwingEntityTableModel) tableModel).conditionModel().enabled(identifier);
-			boolean filterEnabled = tableModel.filterModel().enabled(identifier);
+			boolean conditionEnabled = ((SwingEntityTableModel) tableModel).conditionModel().enabled(identifier).get();
+			boolean filterEnabled = tableModel.filterModel().enabled(identifier).get();
 			boolean showCondition = conditionEnabled || filterEnabled;
 			if (showCondition) {
 				return backgroundColorShaded(row, conditionEnabled && filterEnabled, cellBackgroundColor);